Rename CustomAlert
